requirements.txt
torch>=1.3.1
torchvision>=0.4.2
tensorflow>=1.13.0
numpy==1.16.0
terminaltables
tqdm
Pillow
